Drawer, TMP
Перенес Drawer в вышестоящий класс (для подписки удобнее),
сменил TMP input на обычный

+++ Leadearboard/Assets/Scripts/Leader.cs
@@ -1,9 +1,9 @@
 ﻿public class Leader
 {
     public string Name { get; set; }
-    public int Score { get; set; }
+    public float   Score { get; set; }
 
-    public Leader(string name, int score)
+    public Leader(string name, float score)
     {
         Name = name;
         Score = score;
